requirements.txt
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0
matplotlib>=3.8.0
seaborn>=0.13.0
imbalanced-learn>=0.11.0
xgboost>=1.7.6
lightgbm>=4.0.0
catboost>=1.2.0
shap>=0.44.0
lime>=0.2.0.1
pydantic>=1.10.0
streamlit>=1.28.0
torch>=2.1.0
joblib>=1.3.0
tqdm>=4.66.0
